KFSPTS-37938 Move code into vnd sub-path
