+++ src/main/java/edu/cornell/kfs/cemi/vnd/batch/service/impl/CemiRemitToSupplierFileAppenderCsvImpl.java
@@ -1,4 +1,4 @@
-package edu.cornell.kfs.cemi.scm.remitto.batch.service.impl;
+package edu.cornell.kfs.cemi.vnd.batch.service.impl;
 
 import java.io.File;
 import java.io.IOException;
