+++ src/main/java/edu/cornell/kfs/cemi/vnd/dataaccess/impl/CemiRemitToSupplierDaoJdbcImpl.java
@@ -1,5 +1,5 @@
-package edu.cornell.kfs.cemi.scm.remitto.dataaccess.impl;
-import edu.cornell.kfs.cemi.scm.remitto.dataaccess.CemiRemitToSupplierDao;
+package edu.cornell.kfs.cemi.vnd.dataaccess.impl;
+import edu.cornell.kfs.cemi.vnd.dataaccess.CemiRemitToSupplierDao;
 import edu.cornell.kfs.sys.util.CuSqlQueryPlatformAwareDaoBaseJdbc;
 
 public class CemiRemitToSupplierDaoJdbcImpl extends CuSqlQueryPlatformAwareDaoBaseJdbc implements CemiRemitToSupplierDao {
